Move BPM to global menu, redirection to layer menu
I keep going to the wrong menu for these options because they're placed so
weirdly. The placement of BPM in the layer menu is particularly strange
because it only shows up with the internal clock and all other clock
related options are in the global menu.

The moving output alias options to the layer menu is debatable, but at
least in the way I use it, I tend to think of it more like a property of
the layer ("where does this melody get its gates from?") than a global
setting.