Change probability to mean mute probability
This changes step probability to be more of a continuous setting, like
retrig.

Under the previous definition of step probability, setting it to zero
meant 'always play', but increasing the value to one immediately jumped
to the lowest actual probability of playing the step.

With this change, the definition is inverted and now means 'mute
probability'. Zero still means the step will always play. Increasing the
value makes triggering the gate less likely. The rationale for this is
that it allows dialing in the value in a more natural way.