Improve step hold scheduling
This fixes a bunch of issues with step holding. The previous
implementation worked by increasing next_step_grid_time according to the
hold duration, effectively delaying scheduling of the next step. The
unfortunate side effect of this was that page changes and play position
changes would also be delayed.

The new approach implements hold by maintaining a hold counter that
counts down to zero while the step is held. This makes it possible to
cancel hold early when switching to a different page.