Make cursor-avoidance of popup automatic
This changes removes all explicit calls of g_popup.avoid() in the action
handler code and simply calls it after every action. There were a couple
of cases already where the .avoid() call was missing. With this change
in place, you won't have to think about the cursor-avoidance ever again
because it's just automatic.